+++ src/game/tests.rs
@@ -386,6 +386,7 @@ mod vintage_cube_surveil_land;
 mod vintage_cube_sword_of_the_meek;
 mod vintage_cube_tamiyo;
 mod vintage_cube_tendrils;
+mod vintage_cube_tersa_lightshatter;
 mod vintage_cube_thoughtseize;
 mod vintage_cube_time_spiral;
 mod vintage_cube_titania;